requirements.txt
flask
requests
mysql-connector-python
python-dotenv
gunicorn
flask-cors
fastapi
uvicorn
pymongo
python-dotenv
psycopg2-binary
sqlalchemy